sign의 성별 코드 모듈화 및 코드 정리
