--- src/Components/Sign/SignEmailValidate.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const SignEmailValidate = () => {
-    return <div></div>;
-};
-
-export default SignEmailValidate;
