+++ src/app/(nonAuthLayout)/find_email/page.tsx
@@ -1,10 +1,13 @@
 "use client";
+
 import React from "react";
 import { useState } from "react";
-import { supabase } from "@/util/supabase/clientSupabase";
 
 import SignButton from "@/Components/Sign/SignButton";
 import SignUpItem from "@/Components/Sign/SignUpItem";
+
+import { supabase } from "@/util/supabase/clientSupabase";
+
 const Find_Email = () => {
     const [userEmail, setUserEmail] = useState("");
     const [inputDisabled, setInputDisabled] = useState(false);
