+++ src/app/(nonAuthLayout)/signup/page.tsx
@@ -1,20 +1,22 @@
 "use client";
 
 import { useState, useEffect } from "react";
-import { supabase } from "@/util/supabase/clientSupabase";
-import SignUpItem from "@/Components/Sign/SignUpItem";
-import useInput from "@/hooks/useInput";
 import { useRouter } from "next/navigation";
+import Image from "next/image";
+
+import SignUpItem from "@/Components/Sign/SignUpItem";
+import SignSelectSex from "@/Components/Sign/SignSelectSex";
+import SignPasswordValidate from "@/Components/Sign/SignPasswordValidate";
+import SignUploadBitrthDay from "@/Components/Sign/SignUploadBitrthDay";
+import SignButton from "@/Components/Sign/SignButton";
+import SignPhoneNumber from "@/Components/Sign/SignPhoneNumber";
+
+import useInput from "@/hooks/useInput";
+
+import { supabase } from "@/util/supabase/clientSupabase";
 import { emailValidate } from "@/util/sign/sign_validate";
 import { signUpValidation } from "@/util/sign/signNumber_validation";
-
 import { passwordValidate } from "@/util/sign/sign_validate";
-import SignPasswordValidate from "@/Components/Sign/SignPasswordValidate";
-import SignButton from "@/Components/Sign/SignButton";
-import Image from "next/image";
-import SignUploadBitrthDay from "@/Components/Sign/SignUploadBitrthDay";
-import SignPhoneNumber from "@/Components/Sign/SignPhoneNumber";
-const clickSex = ["남자", "여자"];
 
 const SignUp = () => {
     const [email, onChangeEmailHandler] = useInput();
@@ -43,6 +45,9 @@ const SignUp = () => {
     const [sex, setSex] = useState("");
 
     const router = useRouter();
+    const birthDate = birthYear + birthMonth + birthDay;
+    const phoneNumber = firstNumber + middlePhoneNumber + lastPhoneNumber;
+
     useEffect(() => {
         emailValidate({ email, setEmailRegValid });
         email.length >= 1 ? setEmailError(false) : setEmailError(true);
@@ -57,7 +62,8 @@ const SignUp = () => {
 
     const onChangePassword = (e: React.ChangeEvent<HTMLInputElement>) => setPassword(e.target.value);
 
-    const onClickFindSex = (sex: string) => setSex(sex);
+    const onClickSelectSex = (sex: string) => setSex(sex);
+
     const onClickPhoneNumber = (e: React.ChangeEvent<HTMLSelectElement>) => setFirstNumber(e.target.value);
 
     const onClickBirthYear = (e: React.ChangeEvent<HTMLSelectElement>) => setBirthYear(e.target.value);
@@ -65,7 +71,6 @@ const SignUp = () => {
     const onClickBirthMonth = (e: React.ChangeEvent<HTMLSelectElement>) => setBirthMonth(e.target.value);
 
     const onClickBirthDay = (e: React.ChangeEvent<HTMLSelectElement>) => setBirthDay(e.target.value);
-    const birthDate = birthYear + birthMonth + birthDay;
 
     const onChangeMiddlePhoneNumber = (e: React.ChangeEvent<HTMLSelectElement>) => {
         const { value } = e.target;
@@ -79,8 +84,6 @@ const SignUp = () => {
         setLastPhoneNumber(onlyNumber);
     };
 
-    const phoneNumber = firstNumber + middlePhoneNumber + lastPhoneNumber;
-    console.log(birthDate.length);
     const signUpNewUser = async (e: React.FormEvent<HTMLFormElement>) => {
         e.preventDefault();
         try {
@@ -161,24 +164,8 @@ const SignUp = () => {
                         onClickBirthMonth={onClickBirthMonth}
                         onClickBirthDay={onClickBirthDay}
                     />
+                    <SignSelectSex onClickSelectSex={onClickSelectSex} />
 
-                    <p className="mx-9">성별</p>
-                    <div className="mx-9 mt-[9px] mb-8 h-fit flex flex-row ">
-                        {clickSex.map((item: string, idx: number) => {
-                            return (
-                                <div key={idx} className="flex flex-row my-auto mr-4">
-                                    <label className="mr-1">{item}</label>
-                                    <input
-                                        type="radio"
-                                        name="sex"
-                                        value={item}
-                                        id={item}
-                                        onClick={() => onClickFindSex(item)}
-                                    />
-                                </div>
-                            );
-                        })}
-                    </div>
                     <SignPhoneNumber
                         onClickPhoneNumber={onClickPhoneNumber}
                         onChangeMiddlePhoneNumber={onChangeMiddlePhoneNumber}
